requirements.txt
requests
web3
websocket-client
pysha3
